use larger scrollbar
I noticed that the top few pixels of the scrollbar were not responding to mouse
events. Setting 'overflow: hidden' on the YLCV fixed the problem, as did setting
'display: block' on the YLCV's canvas. I suspect some kind of newline text was overflowing past the YLCV's bounds and capturing the scrollbar's mouse events.