VisualStyle: reform x-label and pointer responsibilities
Now the ChunkView is wholly responsible for reporting x-axis label information from its draw routine. I believe this is a constructive change because x-axis label information is meaningless without considering the stretching technique (which the old x-axis label routines did not accomodate).

In the new system of presenting a ViewStyle, the old pointer coordinate system did not make very much sense. Now, the poitner position is specified in canvas coordinates. Since the ChunkView deals with the canvas in its draw method, this is workable and easier on both the VisualStyle and ContentView.