BarChunkView: drafted justified stretching
Justified stretching works by translating the complete morphing landscape to be at the left, center, or right of the viewport. This is relatively simple to do with _drawRange(), since we can emulate whatever x offset we want.

Note that the _drawStretched() method is not complete yet, since it does not support STRETCH_MODE_ELONGATE. This will require a slightly different implementation, but will still be able to utilize _drawRange().