Dot/CurveChunkView: position modified dots properly.
If an inserting/deleting dot is at the edge, it should not be drawn to be centered in the "bar". Rather, it should be drawn at the edge of the content. This makes insertions/deletions look more natural at the edges. It also makes the spline line align with its corresponding dot during animations, even at the first and last data point.