DotStyle: explicit bottomMargin
Before, no dots were allowed to pass below the bottom of the viewport. Now, they may unless the user of the DotStyle specifically specifies a bottomMargin equal to the radius of the dots.

A side effect of this is that CurveStyle properly handles bottom margins now. Before, it looked really awful when a dot with a 0-value was inserted.