YLCV: propagate pointer events from BufferedView
This will trigger pointer events on the underlying ChunkView when the user moves, leaves, or clicks with their mouse.

Right now, the BufferView is not responsible for handling ChunkView redraw events, but pointer events will probably trigger those. In the future, I may move redraw event handling to the BufferView.