YLCV: show splash screen when resized on ephemeral content
After a canvas resize, our ephemeral content is no longer drawn inside the canvas. Since the ChunkView is gone, we cannot redraw the content. Even if we saved a bitmap of the old content, it would clearly be the wrong size and might require stretching or cropping. The only solution is to jump to the splash screen when the canvas is resized during STATE_EPHEMERAL_CONTENT.