ylcv: drafted skeleton for HeadlessView
The HeadlessView will maintain the state of the YLCV. It is responsible for lazy loading, animations, etc. It does not worry itself with drawing or buffering (as that is the BufferedView's job). It will emit 'change' events whenever its status changes. The YLCV should be able to use the HeadlessView's state as a cue for what to draw in the BufferedView.

Right now, I have all the pertinent event handlers and methods set up. All that is left is to actually *handle* all of the events and compute the states of everything. This way, I can see the big picture, looking at all the potential events that might affect the HeadlessView.