ylcv: drafted HeadlessView._updateYLabels()
This function recomputes the current y-axis labels. This is a fairly isolated method, since it will not change any part of the steady state besides the yLabels field itself. In general, the current y-axis labels are pretty isolated because they do not influence the width of the content--they are an independent variable.