ylcv: drafted HeadlessView._updateLeftmostLabels()
The leftmost labels affect both the scroll offset and the total width. When the leftmost labels are updated, we try to preserve the visible portion of data as w
ell as possible. This makes it possible to manipulate the scroll state before re
computing the leftmost labels, since recomputing the leftmost labels will have n
o affect on the visible region.