ylcv: draft and use HeadlessView._satisfyNeeds()
This method fetches new chunks from the DataSource. It makes sure to enforce intuitive loading behavior: new data is loaded, and changes in needs are met as long as a fetch was already underway; however, if a load fails, resizing the graph or scrolling will not continuously trigger new reloads.