ylcv: implemented "zigzag"
The "zigzag" is the graphic which indicates the edge of the loaded content. I did a bit of geometry to figure out exactly what to draw in the canvas to make the zigzag take up a certain width (breadth = width - lineWidth*sqrt(2)).

Note that the zipzag is used to clip the content. This will come in handy when we draw the horizontal lines for the y-axis labels.