fixed BarChunkView stretch bug
Sometimes, the BarChunkView would think it was being stretched when it wasn't. If all the points were at least partially visible, it thought it was being stretched. This is a false assumption: if the edge points are partially hidden, the view is still not being stretched.

This caused a very unusual problem in the YLCV. You could cause the problem by deleting just enough data points that the content needed to scroll a few pixels but should not stretch.