ylcv: fix Labels.getWidth() bug
The Labels recomputed their width every time you asked for it. First, this was a waste of CPU (although it was only a small amount). More importantly, this broke Labels.transitionFrame(). The problem was that any manually-determined width was being overridden every getWidth() call, since it was recomputed.

I noticed this problem while modifying data points in such a way that the y-axis labels changed widths. The width change was jumping rather than gradually changing.