blurb prototype: hide arrow on blurb in edge cases
Sometimes, there is no logical way to make the blurb's arrow point at a point. For instance, if the point is near the bottom left edge of the viewport, it is difficult to "point" to it with a blurb. Right now, the solution is to show a blurb without an arrow, resembling a regular tooltip.