don't show blurbs for content hidden by the y-axis labels
Before, you could hover over a bar partially or completely hidden by the y-axis labels and see a blurb for it. Now, blurbs only show if your pointer is within the viewport. In addition, blurbs don't always stick out of the center of bars anymore--if the bar is almost offscreen, the blurbs don't go off anymore.