refactor chunk views with BarDrawParams
An instance of this class is passed around during each draw. It can perform translations to/from canvas coordinates, and it deep-copies its arguments for use in the BarXMarkers. This change simplified a whole lot of stuff, because there is a more explicit way to translate between viewport and canvas coordinates.