ylcv: avoid jumps for off-screen insertions
Before this commit, if you scrolled to the left a few hundred pixels, then inserted a data point at the end of the graph, everything would jump by one bar's width. Now, the HeadlessView proactively scrolls so that the data the user is looking at doesn't change when an insertion is made out of bounds.